исправил тип свойства класса

+++ app/src/main/java/ru/geekbrains/filmserach/model/dto/Rating.kt
@@ -7,8 +7,8 @@ data class Rating (
   @SerializedName("_id"                ) var Id                 : String? = null,
   @SerializedName("kp"                 ) var kp                 : Double? = null,
   @SerializedName("imdb"               ) var imdb               : Double? = null,
-  @SerializedName("filmCritics"        ) var filmCritics        : Int?    = null,
-  @SerializedName("russianFilmCritics" ) var russianFilmCritics : Int?    = null,
+  @SerializedName("filmCritics"        ) var filmCritics        : Double? = null,
+  @SerializedName("russianFilmCritics" ) var russianFilmCritics : Double? = null,
   @SerializedName("await"              ) var await              : Int?    = null
 
 )
